fix change password empty update and add rating endpoint

+++ lib/app/data/model/destination_details.dart
@@ -8,7 +8,7 @@ class DestinationDetails {
   String? hostPhone;
   String? destinationAddress;
   String? destinationImg;
-  int? destinationEvaluation;
+  double? destinationEvaluation;
   int? isFavorite;
 
   DestinationDetails(
