+++ lib/app/data/model/destination_model.dart
@@ -31,7 +31,7 @@ class Destination {
   String? destinationAddress;
   String? destinationImg;
   bool? destinationStauts;
-  int? destinationEvaluation;
+  double? destinationEvaluation;
   // bool? destinationFavorite;
 
   Destination({
